requirements.txt
# ----------------------------------------------- Django ---------------------------------------------------------------
Django==5.1.1
python-decouple==3.8
djangorestframework==3.15.2
drf-yasg==1.21.7
ipython==8.27.0

# ----------------------------------------------- Testing --------------------------------------------------------------
rstr==3.2.2
pytest==8.3.3

# ----------------------------------------------- Database -------------------------------------------------------------
mysql-client==0.0.1
mysqlclient==2.2.4
django-redis==5.4.0
redis==5.0.8
factory_boy==3.3.1
Faker==28.4.1

# ---------------------------------------------- Formatting ------------------------------------------------------------
black==23.1.0
flake8==6.0.0
isort==5.12.0